// src/shared/enums/rnc-enums/rnc-epic-transitions-id.ts
export enum RNCEpicTransitionsId {
  ACIONAR_QUALIDADE = '91',
  SAC_ACIONA_QUALIDADE = '151',
  ACIONAR_PCP = '131',
  ACIONAR_SAC = '141',
  ACIONAR_SETOR_DE_ORIGEM = '11',
  VOLTAR_SETOR_EMITENTE = '101',
  VALIDAR_RELATORIO = '31',
  INVALIDAR_RELATORIO = '21',
  REANALISAR = '41',
  ACIONAR_DIRETORIA = '51',
  APROVAR_RELATORIO_EM_PRIMEIRA_INSTANCIA = '71',
  APROVAR_RELATORIO_EM_SEGUNDA_INSTANCIA = '61',
  DEVOLVER_PARA_APROVADA = '191',
  DIRETORIA_REJEITA = '171',
  FINALIZAR_TAREFA = '181',
  CANCELAR_RELATORIO = '111',
}